wip: first try at an "opaque" ffi value
the idea is that we can't/won't/don't want to wrap all values, for many
values it's ok to just get the raw return value from a C function and
only use it as an argument to other C functions.

this doesn't work yet, though, e.g. the following fails with a
segmentation fault, even in when running `make run_interactive`:

    user => (def popen (ffi-fn libc "popen" [String String] FFIOpaque))
    ...
    user => (popen "README.md" "r")
    Segmentation fault.